# docs/swagger.yaml
definitions:
  client.Result:
    properties:
      error:
        type: string
      messages:
        items:
          $ref: '#/definitions/github_com_influxdata_influxdb1-client_v2.Message'
        type: array
      series:
        items:
          $ref: '#/definitions/models.Row'
        type: array
      statement_id:
        type: integer
    type: object
  github_com_influxdata_influxdb1-client_v2.Message:
    properties:
      level:
        type: string
      text:
        type: string
    type: object
...
  models.Row:
    properties:
      columns:
        items:
          type: string
        type: array
      name:
        type: string
      partial:
        type: boolean
      tags:
        additionalProperties:
          type: string
        type: object
      values:
        items:
          items: {}
          type: array
        type: array
    type: object
info:
  contact: {}
  description: Provides HTTP-API to request current and historical connection log.
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Connection Log API
  version: '{version}'
paths:
...
  /intern/history/device/{duration}:
    post:
      consumes:
      - application/json
      parameters:
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      - description: duration in influxdb format https://docs.influxdata.com/influxdb/v1.5/query_language/spec/#durations
        in: path
        name: duration
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: result
          schema:
            items:
              $ref: '#/definitions/client.Result'
            type: array
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern get devices history
      tags:
      - Old api
  /intern/history/gateway/{duration}:
    post:
      consumes:
      - application/json
      parameters:
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      - description: duration in influxdb format https://docs.influxdata.com/influxdb/v1.5/query_language/spec/#durations
        in: path
        name: duration
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: result
          schema:
            items:
              $ref: '#/definitions/client.Result'
            type: array
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern get gateways history
      tags:
      - Old api
  /intern/logedge/device/{duration}:
    post:
      consumes:
      - application/json
      parameters:
      - description: duration in influxdb format https://docs.influxdata.com/influxdb/v1.5/query_language/spec/#durations
        in: path
        name: duration
        required: true
        type: string
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                items: {}
                type: array
              type: array
            type: object
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern get devices log edge
      tags:
      - Old api
  /intern/logedge/gateway/{duration}:
    post:
      consumes:
      - application/json
      parameters:
      - description: duration in influxdb format https://docs.influxdata.com/influxdb/v1.5/query_language/spec/#durations
        in: path
        name: duration
        required: true
        type: string
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                items: {}
                type: array
              type: array
            type: object
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern get gateways log edge
      tags:
      - Old api
  /intern/logstarts/device:
    post:
      consumes:
      - application/json
      parameters:
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: unix timestamps mapped to IDs
          schema:
            additionalProperties:
              format: float64
              type: number
            type: object
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern get devices log start
      tags:
      - Old api
  /intern/logstarts/gateway:
    post:
      consumes:
      - application/json
      parameters:
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: unix timestamps mapped to IDs
          schema:
            additionalProperties:
              format: float64
              type: number
            type: object
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern get gateways log start
      tags:
      - Old api
  /intern/state/device/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: states mapped to IDs
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern check device online states
      tags:
      - Old api
  /intern/state/gateway/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: states mapped to IDs
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Intern check gateway online states
      tags:
      - Old api
  /state/device/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: list of IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: states mapped to IDs
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: error message
          schema:
            type: string
        "401":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Check device online states
      tags:
      - Old api
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"